// apps/web-v4/src/app/models/apiResponse.ts
/*               "Copyright 2020 Infosys Ltd.
               Use of this source code is governed by GPL v3 license that can be found in the LICENSE file or at https://opensource.org/licenses/GPL-3.0
               This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License version 3"*/
export interface IApiResponse<T> {
  id: string;
  ver: string;
  ts: string;
  params: {
    resmsgid: string
    msgid: string
    err: any
    status: string
    errmsg: string
  };
  responseCode: string;
  result: T;
}

export interface IApiResponseResultResponse<T> {
  response: T;
}